asd
asd

+++ index.js
@@ -393,6 +393,20 @@ bot.on('message', async message => {
             message.member.roles.remove(chance)
             return
         }
+        if(message.channel.id == '820645410175320085') return
+        await message.delete()
+        return message.channel.send('Du darfst hier keine Links reinschicken!')
+    }
+})
+
+bot.on('message', async message => {
+    let link = ["discord.gg", "discord.com/invite", "discordapp.com/invite", "https://", "http://"]
+
+    if (link.some(word => message.content.toLowerCase().includes(word))) {
+        if (message.member.roles.cache.find(r => r.id == chance)) {
+            message.member.roles.remove(chance)
+            return
+        }
         if(message.channel.id == '811517449718988801') return
         await message.delete()
         return message.channel.send('Du darfst hier keine Links reinschicken!')
